% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSSPCA.R
\name{iSSPCA}
\alias{iSSPCA}
\title{This function decomposes a matrix using singular value decomposition.
The resulting eigenvector corresponding to the loadings in then shrunk
using an L1 penalty.}
\usage{
iSSPCA(
  X,
  Y,
  xts,
  sumabsv = 4,
  niter = 20,
  npc = 1,
  orth = TRUE,
  trace = FALSE,
  v = NULL,
  center = TRUE,
  cnames = NULL,
  vpos = FALSE,
  vneg = FALSE,
  compute.pve = TRUE,
  Ykernel = c("linear", "delta"),
  decomposition.type = c("default", "RSpectra")
)
}
\arguments{
\item{X}{The training feature space}

\item{Y}{The response vector}

\item{xts}{The testing feature space}

\item{sumabsv}{sparseness penalty}

\item{npc}{number of principal components}
}
\description{
This function decomposes a matrix using singular value decomposition.
The resulting eigenvector corresponding to the loadings in then shrunk
using an L1 penalty.
}
\examples{
SMD(x, sumabsu, sumabsv, niter=20,trace=TRUE, v, upos, uneg, vpos, vneg)
}
\keyword{function}
\keyword{parent}
