% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respkernel.R
\name{respkernel}
\alias{respkernel}
\title{A function kernelizes the response data}
\usage{
respkernel(Y, kernel, nobs)
}
\arguments{
\item{Y}{Vector of resposne data}

\item{kernel}{Desired kernel for the response data}

\item{nobs}{number of observations in the dataset}
}
\description{
A function kernelizes the response data
}
\examples{
respkernel(Y, kernel, nobs)
}
\keyword{response}
