% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPCA.R
\name{SPCA}
\alias{SPCA}
\title{This function decomposes a matrix using singular value decomposition.
The resulting eigenvector corresponding to the loadings in then shrunk
using an L1 penalty.}
\usage{
SPCA(X, Y, npc, decomposition.type = c("default", "RSpectra"), nobs, p, L, H)
}
\arguments{
\item{X}{This contains the training feature space}

\item{Y}{This contains the response vector}

\item{npc}{This is the number principal components desired}

\item{decomposition.type}{Determines which decomposition method is used.}

\item{nobs}{This contains the number of observations}

\item{p}{This contains the number of predictors}

\item{L}{This is the response kernel matrix}

\item{H}{This is the centering matrix}
}
\description{
This function decomposes a matrix using singular value decomposition.
The resulting eigenvector corresponding to the loadings in then shrunk
using an L1 penalty.
}
\examples{
SMD(x, sumabsu, sumabsv, niter=20,trace=TRUE, v, upos, uneg, vpos, vneg)

}
\keyword{PCA}
\keyword{supervised}
